update table actions infer
