+++ ymir/web/typings.d.ts
@@ -8,3 +8,19 @@ declare module '*.svg' {
   const url: string;
   export default url;
 }
+// components table actions
+type Action = {
+  key: string, 
+  label: string, 
+  onclick?: Function,
+  icon?: string, 
+  link?: string, 
+  target?: string,
+  disabled?: boolean,
+  hidden?: Function,
+}
+
+type ComponentsTableActionsProps = {
+  actions: Action[],
+  showCount?: number,
+}
